# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI MVP API
  version: 0.2.0
  description: BuildGraph 3주 MVP 공통 API 계약. 상세 기준은 docs/API_CONTRACT.md를 따른다.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/parts:
    get:
      summary: 부품 목록
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/PartCategory"
        - in: query
          name: q
          schema:
            type: string
          description: 부품명, 제조사, attributes JSON 검색어
        - in: query
          name: manufacturer
          schema:
            type: string
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/GenericStatus"
        - in: query
          name: minPrice
          schema:
            type: integer
            minimum: 0
        - in: query
          name: maxPrice
          schema:
            type: integer
            minimum: 0
        - in: query
          name: sort
          schema:
            type: string
            enum: [category, price_asc, price_desc, name]
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 부품 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartPage"
  /api/parts/{id}:
    get:
      summary: 부품 상세
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: 부품 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartDto"
  /api/parts/{id}/price-history:
    get:
      summary: 부품 가격 이력
      description: 저장된 price_snapshots를 기준으로 상품별 가격변동 추이를 조회한다. 외부 검색 API를 실시간 호출하지 않는다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PublicId"
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
            maximum: 3650
            default: 90
        - in: query
          name: source
          schema:
            type: string
          description: "특정 가격 출처만 조회할 때 사용한다. 예: NAVER_SHOPPING_SEARCH"
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 120
      responses:
        "200":
          description: 부품 가격 이력
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartPriceHistoryResponse"
  /api/quote-drafts/current:
    get:
      summary: 현재 로그인 사용자의 수동 견적초안
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 현재 견적초안. active draft가 없으면 status=EMPTY로 반환한다.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDraftDto"
  /api/quote-drafts/current/items/{partId}:
    put:
      summary: 견적초안에 부품 담기 또는 같은 카테고리 교체
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: partId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteDraftItemUpsertRequest"
      responses:
        "200":
          description: 갱신된 견적초안
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDraftDto"
    patch:
      summary: 견적초안 부품 수량 변경
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: partId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteDraftItemUpsertRequest"
      responses:
        "200":
          description: 갱신된 견적초안
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDraftDto"
    delete:
      summary: 견적초안에서 부품 제거
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: partId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 갱신된 견적초안
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteDraftDto"
  /api/price-alerts:
    get:
      summary: 목표가 알림 목록
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 목표가 알림 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlertPage"
    post:
      summary: 목표가 알림 등록
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceAlertCreateRequest"
      responses:
        "201":
          description: 생성된 알림
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlertDto"
...
  /api/admin/parts/external-offers/refresh:
    post:
      summary: 외부 상품 캐시 갱신
      description: 네이버 쇼핑 검색 API를 호출해 part_external_offers 캐시를 갱신한다. 사용자 부품 조회 API는 이 캐시만 읽는다.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/PartCategory"
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: force
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: 갱신 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalOfferRefreshResponse"
  /api/admin/parts/catalog/refresh:
    post:
      summary: 내부 자산 카탈로그 후보 갱신
      description: 카테고리별 query pack으로 외부 검색 API를 호출해 part_catalog_candidates를 채우고, publish=true일 때 parts에 게시한다. 사용자 부품 조회 API는 외부 API를 직접 호출하지 않는다.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          required: true
          schema:
            $ref: "#/components/schemas/PartCategory"
        - in: query
          name: limitPerQuery
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 3
        - in: query
          name: publish
          schema:
            type: boolean
            default: false
        - in: query
          name: q
          schema:
            type: string
          description: 단일 검색어 테스트용. 생략하면 category별 query pack을 사용한다.
      responses:
        "200":
          description: 내부 자산 갱신 결과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartCatalogRefreshResponse"
  /api/tools/compatibility/check:
    post:
      summary: 호환성 Tool check
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ToolCheckRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/ToolCheckSuccess"
  /api/tools/power/check:
    post:
      summary: 전력 Tool check
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ToolCheckRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/ToolCheckSuccess"
  /api/tools/size/check:
    post:
      summary: 규격 Tool check
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ToolCheckRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/ToolCheckSuccess"
  /api/tools/performance/check:
    post:
      summary: 성능 Tool check
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ToolCheckRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/ToolCheckSuccess"
  /api/tools/price/check:
    post:
      summary: 가격 Tool check
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/ToolCheckRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/ToolCheckSuccess"
...
  /api/rag/search:
    get:
      summary: RAG 근거 검색
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: RAG 근거 후보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagEvidencePage"
  /api/rag/evidence/{id}:
    get:
      summary: RAG 근거 상세
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: RAG 근거 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagEvidenceDto"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PublicId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  requestBodies:
    ToolCheckRequestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ToolCheckRequest"
  responses:
    ToolCheckSuccess:
      description: Tool check 결과
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ToolCheckResponse"
  schemas:
...
    GenericStatus:
      type: string
      enum: [ACTIVE, INACTIVE, DISCONTINUED]
    ToolStatus:
      type: string
      enum: [PASS, WARN, FAIL]
    Confidence:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    PartCategory:
      type: string
      enum: [CPU, GPU, RAM, MOTHERBOARD, STORAGE, PSU, CASE, COOLER]
...
    PriceAlertStatus:
      type: string
      enum: [ACTIVE, TRIGGERED, DISABLED]
...
    ToolEvidenceDto:
      type: object
      required: [source_id, summary]
      properties:
        source_id:
          type: string
        summary:
          type: string
...
    PartDto:
      type: object
      required: [id, category, name, price, status, attributes]
      properties:
        id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/PartCategory"
        name:
          type: string
        manufacturer:
          type: string
        price:
          type: integer
        status:
          $ref: "#/components/schemas/GenericStatus"
        attributes:
          type: object
          additionalProperties: true
        benchmarkSummary:
          type: object
          additionalProperties: true
        latestPriceSource:
          type: string
          nullable: true
        latestPriceCollectedAt:
          type: string
          format: date-time
          nullable: true
        externalOffer:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PartExternalOfferDto"
    PartExternalOfferDto:
      type: object
      properties:
        title:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        supplierName:
          type: string
          nullable: true
        offerUrl:
          type: string
          nullable: true
        lowPrice:
          type: integer
          nullable: true
        source:
          type: string
        refreshedAt:
          type: string
          format: date-time
          nullable: true
    QuoteDraftItemUpsertRequest:
      type: object
      description: CPU/GPU/MOTHERBOARD/PSU/CASE/COOLER are single-item categories. RAM and STORAGE can hold multiple distinct parts.
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 9
          default: 1
          description: Single-item categories accept only 1. RAM and STORAGE accept 1 to 9.
    QuoteDraftDto:
      type: object
      required: [status, name, items, totalPrice, itemCount]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [EMPTY, ACTIVE, ORDERED, ARCHIVED]
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteDraftItemDto"
        totalPrice:
          type: integer
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
    QuoteDraftItemDto:
      type: object
      required: [id, partId, category, name, quantity, unitPriceAtAdd, currentPrice, lineTotal]
      properties:
        id:
          type: string
          format: uuid
        partId:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/PartCategory"
        name:
          type: string
        manufacturer:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPriceAtAdd:
          type: integer
        currentPrice:
          type: integer
        lineTotal:
          type: integer
        attributes:
          type: object
          additionalProperties: true
        externalOffer:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PartExternalOfferDto"
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
    PartPriceHistoryResponse:
      type: object
      required: [partId, partName, currentPrice, days, items, summary]
      properties:
        partId:
          type: string
          format: uuid
        partName:
          type: string
        currentPrice:
          type: integer
        days:
          type: integer
        source:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PriceHistoryPoint"
        summary:
          $ref: "#/components/schemas/PriceHistorySummary"
    PriceHistoryPoint:
      type: object
      required: [price, source, collectedAt]
      properties:
        price:
          type: integer
        source:
          type: string
        collectedAt:
          type: string
          format: date-time
    PriceHistorySummary:
      type: object
      required: [sampleCount, currentPrice, minPrice, maxPrice, firstPrice, lastPrice, changeAmount, changeRatePercent]
      properties:
        sampleCount:
          type: integer
        currentPrice:
          type: integer
        minPrice:
          type: integer
        maxPrice:
          type: integer
        firstPrice:
          type: integer
        lastPrice:
          type: integer
        changeAmount:
          type: integer
        changeRatePercent:
          type: number
    PriceAlertCreateRequest:
      type: object
      required: [partId, targetPrice]
      properties:
        partId:
          type: string
          format: uuid
        targetPrice:
          type: integer
    PriceAlertDto:
      type: object
      required: [partId, partName, targetPrice, currentPrice, status]
      properties:
        partId:
          type: string
          format: uuid
        partName:
          type: string
        targetPrice:
          type: integer
        currentPrice:
          type: integer
        status:
          $ref: "#/components/schemas/PriceAlertStatus"
        createdAt:
          type: string
          format: date-time
...
    ExternalOfferRefreshResponse:
      type: object
      required: [configured, limit, attempted, updated, skipped, failed]
      properties:
        configured:
          type: boolean
        category:
          $ref: "#/components/schemas/PartCategory"
          nullable: true
        limit:
          type: integer
        force:
          type: boolean
        attempted:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        message:
          type: string
    PartCatalogRefreshResponse:
      type: object
      required: [configured, category, queryCount, limitPerQuery, attempted, discovered, published, failed]
      properties:
        configured:
          type: boolean
        jobId:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/PartCategory"
        queryCount:
          type: integer
        limitPerQuery:
          type: integer
        publish:
          type: boolean
        attempted:
          type: integer
        discovered:
          type: integer
        published:
          type: integer
        failed:
          type: integer
        errorSummary:
          type: string
          nullable: true
        message:
          type: string
    ToolCheckRequest:
      type: object
      required: [partIds, context]
      properties:
        buildId:
          type: string
          format: uuid
        partIds:
          type: array
          items:
            type: string
            format: uuid
        context:
          type: object
          additionalProperties: true
    ToolCheckResponse:
      type: object
      required: [tool, status, score, confidence, summary, warnings, evidence, details]
      properties:
        tool:
          type: string
        status:
          $ref: "#/components/schemas/ToolStatus"
        score:
          type: number
          format: double
        confidence:
          $ref: "#/components/schemas/Confidence"
        summary:
          type: string
        warnings:
          type: array
          items:
            type: string
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/ToolEvidenceDto"
        details:
          type: object
          additionalProperties: true
...
    RagEvidenceDto:
      type: object
      required: [id, sourceId, summary]
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
        summary:
          type: string
        score:
          type: number
        metadata:
          type: object
          additionalProperties: true
...
    PartPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PartDto"
    PriceAlertPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PriceAlertDto"
...
    RagEvidencePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/RagEvidenceDto"
...
    PageMeta:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer